PICARD-2454: Stop metadatabox updates while moving files
